WIP: factoring apart tests for py2/3 async adapter support
